Highlight tracks that failed when saving
Normally, it shouldn't be possible for this to happen as each track is
validated before the playlist is saved. This can happen, however, if
tracks are deleted from the underlying source before saving. If that
happens, give the user some indication of what actually went wrong by
highlighting the tracks and adding a tooltip.

Fixes #65.